Create a way where we can convert the found graph path into something more human readable.
For that we refactor the path controller since it's not really needed anymore. There's still stuff to do, but this is just to make it work as a first iteration. We've for instance removed everything related to threads as we think it won't be needed and thus there's no need to overcomplicacte it